Project page layout changes
I've changed a lot of things; set up some more divs, most importantly moving class="project-section" div to one right above the div with the content in order to set it at a negative margin and scroll without having huge gaps of white on my page. Secondary, prototype on Vela is now a flex column with About section; still trying to figure out how to get it to auto-resize.